Regular upgrade[3.1.1-SNAPSHOT]

+++ albedo-plugins/albedo-data-mybatis/src/main/java/com/albedo/java/common/persistence/service/TreeService.java
@@ -21,7 +21,7 @@ public interface TreeService<T extends TreeEntity, D extends TreeDto> extends Da
 	 * @param parentId
 	 * @return
 	 */
-	Integer countByParentId(String parentId);
+	Long countByParentId(String parentId);
 
 	/**
 	 * getTreeWrapper
